feat: add copy conference info button to main page
