+++ src/app/conference/page.tsx
@@ -14,12 +14,9 @@
  * limitations under the License.
  */
 
-import InstantMeetingButton from "@/components/InstantMeeting/InstantMeetingButton";
 import { Metadata } from "next";
 import NavigationTopBar from "@/components/opendesk/NavigationTopBar/NavigationTopBar";
-import VideoTestButton from "@/components/InstantMeeting/VideoTestButton";
-import {Stack} from "@mui/material";
-import {isVarTrue} from "@/lib/isVarTrue";
+import ConferenceActions from "@/components/conference/ConferenceActions";
 
 export const metadata: Metadata = {
     title: 'openDesk Video Conference',
@@ -27,24 +24,11 @@ export const metadata: Metadata = {
 };
 
 export default function Conference() {
-
-    const envVideoTestEnabled = process.env.NEXT_PUBLIC_VIDEO_TEST_ENABLED;
-    const isVideoTestEnabled = isVarTrue(envVideoTestEnabled);
-
     return (
         <>
             <NavigationTopBar />
             <main className="flex flex-col h-screen max-h-[80vh] items-center justify-center m-4">
-                <Stack className="space-y-6 w-full items-center justify-center">
-                    <Stack className="w-1/4">
-                        <InstantMeetingButton />
-                    </Stack>
-                    { isVideoTestEnabled &&
-                        <Stack>
-                            <VideoTestButton />
-                        </Stack>
-                    }
-                </Stack>
+                <ConferenceActions />
             </main>
         </>
     );
